Add C++ version: 524181 nots
C++ (GNU C++ Compiler, g++ version 12.2.0) has a non-fully
deterministic behaviour, although rather consistent.

Up to the 524181 repetition (I used not, which is a valid C++
alias, but the same behaviour happens with !), it compiles almost
always, when more "not"s are added, it frequently fails with the
error:

  g++: internal compiler error: Segmentation fault signal
  terminated program cc1plus
  Please submit a full bug report, with preprocessed source (by
  using -freport-bug).
  See <file:///usr/share/doc/gcc-12/README.Bugs> for instructions.

It is not fully deterministic, as it is not a hard limit, it may
compile beyond that at times or fail unfrequently before that.

To compile:
  g++ not.cpp -o not
To run:
  ./not